feat: add ClashDisplay font files and update font import structure
