+++ src/fonts/index.ts
@@ -1,28 +1,3 @@
-import localFont from "next/font/local";
+import { clashDisplay } from "./clashDisplay";
 
-export const clashDisplay = localFont({
-  src: [
-    {
-      path: "../fonts/ClashDisplay-Light.ttf",
-      weight: "300",
-      style: "normal",
-    },
-    {
-      path: "../fonts/ClashDisplay-Regular.ttf",
-      weight: "400",
-      style: "normal",
-    },
-    {
-      path: "../fonts/ClashDisplay-Semibold.ttf",
-      weight: "500",
-      style: "normal",
-    },
-    {
-      path: "../fonts/ClashDisplay-Bold.ttf",
-      weight: "800",
-      style: "normal",
-    },
-  ],
-  display: "swap",
-  variable: "--font-clash-display",
-});
+export { clashDisplay };
